requirements.txt
scikit-learn==1.0.2
pandas
streamlit==1.8.1
matplotlib==3.5.1
seaborn==0.11.2
numpy
st-annotated-text
plotly
streamlit-option-menu
streamlit_lottie